requirements.txt
google-adk
python-dotenv
requests
beautifulsoup4